requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
pytest
pytest-asyncio
httpx
authlib
passlib[bcrypt]
python-multipart
email-validator